Refactor sp validator
